requirements.txt
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
decorator==5.1.1
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-json-logger==2.0.2
requests==2.27.1
shellingham==1.4.0
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
typer==0.4.1
urllib3==1.26.9
validators==0.19.0
